requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
factory-boy==3.3.0